+++ Repositories/ProjectRepository.cs
@@ -32,7 +32,7 @@ namespace Repositories
     public interface IProjectRepository : IRepository<Project>
     {
 
-        Task<List<Project>> GetAllWithUsersAsync(
+        Task<List<Project>> GetAllWithUsersAndCollaboratorsAsync(
             int? skip = null,
             int? take = null,
             Expression<Func<Project, object>> orderBy = null,
@@ -173,7 +173,7 @@ namespace Repositories
         /// <param name="orderByAsc">The order direction (True: asc, False: desc)</param>
         /// <param name="highlighted">Filter highlighted projects</param>
         /// <returns>The projects filtered by the parameters</returns>
-        public virtual async Task<List<Project>> GetAllWithUsersAsync(
+        public virtual async Task<List<Project>> GetAllWithUsersAndCollaboratorsAsync(
             int? skip = null,
             int? take = null,
             Expression<Func<Project, object>> orderBy = null,
@@ -184,7 +184,9 @@ namespace Repositories
             IQueryable<Project> queryable = DbSet
                                             .Include(p => p.User)
                                             .Include(p => p.ProjectIcon)
-                                            .Include(p => p.CallToAction);
+                                            .Include(p => p.CallToAction)
+                                            .Include(p => p.Collaborators);
+
             queryable = ApplyFilters(queryable, skip, take, orderBy, orderByAsc, highlighted);
 
             List<Project> projects = await queryable.ToListAsync();