+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Automatically link users to their institution - [#295](https://github.com/DigitalExcellence/dex-backend/issues/295)
 - Added call to actions for projects and call to action options - [312](https://github.com/DigitalExcellence/dex-backend/issues/312)
+- Collaborators are now included on the project overview page - [#317](https://github.com/DigitalExcellence/dex-backend/issues/317)
+
 
 ### Changed